Add files via upload
Use the datasets in modified for plots. The crime_dataset_edu includes all data about crime rate, unemployment, income, and education. crime_dataset is the same but without education, difference is it has a longer time frame. Crimes is just the crime data. edu_finance is the same as crime_dataset_edu but includes the financing spent on education, i made it into another dataset because the data is only from 2008-2016